requirements.txt
PyYAML>=6.0
python-dotenv>=1.0
spotipy>=2.24.0
yt-dlp>=2025.12.0
mutagen>=1.46.0
requests>=2.31.0
ffmpeg-python>=0.3.0
SQLAlchemy>=2.0.0
supabase>=1.6.0
